+++ usr.sbin/httpd/src/modules/proxy/proxy_util.c
@@ -1,58 +1,59 @@
 /* ====================================================================
- * Copyright (c) 1996-1998 The Apache Group.  All rights reserved.
+ * The Apache Software License, Version 1.1
+ *
+ * Copyright (c) 2000 The Apache Software Foundation.  All rights
+ * reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
  *
  * 1. Redistributions of source code must retain the above copyright
- *    notice, this list of conditions and the following disclaimer. 
+ *    notice, this list of conditions and the following disclaimer.
  *
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in
  *    the documentation and/or other materials provided with the
  *    distribution.
  *
- * 3. All advertising materials mentioning features or use of this
- *    software must display the following acknowledgment:
- *    "This product includes software developed by the Apache Group
- *    for use in the Apache HTTP server project (http://www.apache.org/)."
+ * 3. The end-user documentation included with the redistribution,
+ *    if any, must include the following acknowledgment:
+ *       "This product includes software developed by the
+ *        Apache Software Foundation (http://www.apache.org/)."
+ *    Alternately, this acknowledgment may appear in the software itself,
+ *    if and wherever such third-party acknowledgments normally appear.
  *
- * 4. The names "Apache Server" and "Apache Group" must not be used to
- *    endorse or promote products derived from this software without
- *    prior written permission. For written permission, please contact
- *    apache@apache.org.
+ * 4. The names "Apache" and "Apache Software Foundation" must
+ *    not be used to endorse or promote products derived from this
+ *    software without prior written permission. For written
+ *    permission, please contact apache@apache.org.
  *
- * 5. Products derived from this software may not be called "Apache"
- *    nor may "Apache" appear in their names without prior written
- *    permission of the Apache Group.
+ * 5. Products derived from this software may not be called "Apache",
+ *    nor may "Apache" appear in their name, without prior written
+ *    permission of the Apache Software Foundation.
  *
- * 6. Redistributions of any form whatsoever must retain the following
- *    acknowledgment:
- *    "This product includes software developed by the Apache Group
- *    for use in the Apache HTTP server project (http://www.apache.org/)."
- *
- * THIS SOFTWARE IS PROVIDED BY THE APACHE GROUP ``AS IS'' AND ANY
- * EXPRESSED OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
- * PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL THE APACHE GROUP OR
+ * THIS SOFTWARE IS PROVIDED ``AS IS'' AND ANY EXPRESSED OR IMPLIED
+ * WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+ * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED.  IN NO EVENT SHALL THE APACHE SOFTWARE FOUNDATION OR
  * ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT
- * NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
- * HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT,
- * STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED
- * OF THE POSSIBILITY OF SUCH DAMAGE.
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF
+ * USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT
+ * OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
+ * SUCH DAMAGE.
  * ====================================================================
  *
  * This software consists of voluntary contributions made by many
- * individuals on behalf of the Apache Group and was originally based
- * on public domain software written at the National Center for
- * Supercomputing Applications, University of Illinois, Urbana-Champaign.
- * For more information on the Apache Group and the Apache HTTP server
- * project, please see <http://www.apache.org/>.
+ * individuals on behalf of the Apache Software Foundation.  For more
+ * information on the Apache Software Foundation, please see
+ * <http://www.apache.org/>.
  *
+ * Portions of this software are based upon public domain software
+ * originally written at the National Center for Supercomputing Applications,
+ * University of Illinois, Urbana-Champaign.
  */
 
 /* Utility routines for Apache proxy */
@@ -68,7 +69,7 @@ static int proxy_match_ipaddr(struct dirconn_entry *This, request_rec *r);
 static int proxy_match_domainname(struct dirconn_entry *This, request_rec *r);
 static int proxy_match_hostname(struct dirconn_entry *This, request_rec *r);
 static int proxy_match_word(struct dirconn_entry *This, request_rec *r);
-
+static struct per_thread_data *get_per_thread_data(void);
 /* already called in the knowledge that the characters are hex digits */
 int ap_proxy_hex2c(const char *x)
 {
@@ -116,10 +117,10 @@ void ap_proxy_c2hex(int ch, char *x)
 	x[2] = '0' + i;
 #else /*CHARSET_EBCDIC*/
     static const char ntoa[] = { "0123456789ABCDEF" };
-    ch &= 0xFF;
+    ch = os_toascii[ch & 0xFF];
     x[0] = '%';
-    x[1] = ntoa[(os_toascii[ch]>>4)&0x0F];
-    x[2] = ntoa[os_toascii[ch]&0x0F];
+    x[1] = ntoa[(ch>>4)&0x0F];
+    x[2] = ntoa[ch&0x0F];
     x[3] = '\0';
 #endif /*CHARSET_EBCDIC*/
 }
@@ -134,8 +135,8 @@ void ap_proxy_c2hex(int ch, char *x)
  * and encodes those which must be encoded, and does not touch
  * those which must not be touched.
  */
-char *
-     ap_proxy_canonenc(pool *p, const char *x, int len, enum enctype t, int isenc)
+char *ap_proxy_canonenc(pool *p, const char *x, int len, enum enctype t,
+			enum proxyreqtype isenc)
 {
     int i, j, ch;
     char *y;
@@ -177,8 +178,8 @@ char *
 	    continue;
 	}
 /* decode it if not already done */
-	if (isenc && ch == '%') {
-	    if (!isxdigit(x[i + 1]) || !isxdigit(x[i + 2]))
+	if (isenc != NOT_PROXY && ch == '%') {
+	    if (!ap_isxdigit(x[i + 1]) || !ap_isxdigit(x[i + 2]))
 		return NULL;
 	    ch = ap_proxy_hex2c(&x[i + 1]);
 	    i += 2;
@@ -239,12 +240,12 @@ char *
 	strp = strchr(user, ':');
 	if (strp != NULL) {
 	    *strp = '\0';
-	    password = ap_proxy_canonenc(p, strp + 1, strlen(strp + 1), enc_user, 1);
+	    password = ap_proxy_canonenc(p, strp + 1, strlen(strp + 1), enc_user, STD_PROXY);
 	    if (password == NULL)
 		return "Bad %-escape in URL (password)";
 	}
 
-	user = ap_proxy_canonenc(p, user, strlen(user), enc_user, 1);
+	user = ap_proxy_canonenc(p, user, strlen(user), enc_user, STD_PROXY);
 	if (user == NULL)
 	    return "Bad %-escape in URL (username)";
     }
@@ -263,11 +264,14 @@ char *
 	    if (!ap_isdigit(strp[i]))
 		break;
 
-	if (i == 0 || strp[i] != '\0')
+	/* if (i == 0) the no port was given; keep default */
+	if (strp[i] != '\0') {
 	    return "Bad port number in URL";
-	*port = atoi(strp);
-	if (*port > 65535)
-	    return "Port number in URL > 65535";
+	} else if (i > 0) {
+	    *port = atoi(strp);
+	    if (*port > 65535)
+		return "Port number in URL > 65535";
+	}
     }
     ap_str_tolower(host);		/* DNS names are case-insensitive */
     if (*host == '\0')
@@ -277,7 +281,7 @@ char *
 	if (!ap_isdigit(host[i]) && host[i] != '.')
 	    break;
     /* must be an IP address */
-#ifdef WIN32
+#if defined(WIN32) || defined(NETWARE) || defined(TPF) || defined(BEOS)
     if (host[i] == '\0' && (inet_addr(host) == -1))
 #else
     if (host[i] == '\0' && (ap_inet_addr(host) == -1 || inet_network(host) == -1))
@@ -488,15 +492,15 @@ table *ap_proxy_read_headers(request_rec *r, char *buffer, int size, BUFF *f)
 
 long int ap_proxy_send_fb(BUFF *f, request_rec *r, cache_req *c)
 {
-    int  ok = 1;
+    int  ok;
     char buf[IOBUFSIZE];
-    long total_bytes_rcv;
+    long total_bytes_rcvd;
     register int n, o, w;
     conn_rec *con = r->connection;
-    int alt_to = 1;
+    int alternate_timeouts = 1;	/* 1 if we alternate between soft & hard timeouts */
 
-    total_bytes_rcv = 0;
-    if (c)
+    total_bytes_rcvd = 0;
+    if (c != NULL)
         c->written = 0;
 
 #ifdef CHARSET_EBCDIC
@@ -514,84 +518,90 @@ long int ap_proxy_send_fb(BUFF *f, request_rec *r, cache_req *c)
 
     ap_kill_timeout(r);
 
-#ifdef WIN32
+#if defined(WIN32) || defined(TPF) || defined(NETWARE)
     /* works fine under win32, so leave it */
     ap_hard_timeout("proxy send body", r);
-    alt_to = 0;
+    alternate_timeouts = 0;
 #else
     /* CHECKME! Since hard_timeout won't work in unix on sends with partial
      * cache completion, we have to alternate between hard_timeout
      * for reads, and soft_timeout for send.  This is because we need
      * to get a return from ap_bwrite to be able to continue caching.
      * BUT, if we *can't* continue anyway, just use hard_timeout.
+     * (Also, if no cache file is written, use hard timeouts)
      */
 
-    if (c) {
-        if (c->len <= 0 || c->cache_completion == 1) {
-            ap_hard_timeout("proxy send body", r);
-            alt_to = 0;
-        }
-    } else {
+    if (c == NULL || c->len <= 0 || c->cache_completion == 1.0) {
         ap_hard_timeout("proxy send body", r);
-        alt_to = 0;
+        alternate_timeouts = 0;
     }
 #endif
 
-    while (ok) {
-        if (alt_to)
-            ap_hard_timeout("proxy send body", r);
+    /* Loop and ap_bread() while we can successfully read and write,
+     * or (after the client aborted) while we can successfully
+     * read and finish the configured cache_completion.
+     */
+    for (ok = 1; ok; ) {
+        if (alternate_timeouts)
+            ap_hard_timeout("proxy recv body from upstream server", r);
 
 	/* Read block from server */
 	n = ap_bread(f, buf, IOBUFSIZE);
 
-        if (alt_to)
+        if (alternate_timeouts)
             ap_kill_timeout(r);
         else
             ap_reset_timeout(r);
 
 	if (n == -1) {		/* input error */
-	    if (c != NULL)
+	    if (c != NULL) {
+		ap_log_rerror(APLOG_MARK, APLOG_ERR, c->req,
+		    "proxy: error reading from %s", c->url);
 		c = ap_proxy_cache_error(c);
+	    }
 	    break;
 	}
 	if (n == 0)
 	    break;		/* EOF */
 	o = 0;
-	total_bytes_rcv += n;
+	total_bytes_rcvd += n;
 
 	/* Write to cache first. */
+	/*@@@ XXX FIXME: Assuming that writing the cache file won't time out?!!? */
         if (c != NULL && c->fp != NULL) {
             if (ap_bwrite(c->fp, &buf[0], n) != n) {
-                c = ap_proxy_cache_error(c);
+                ap_log_rerror(APLOG_MARK, APLOG_ERR, c->req,
+		    "proxy: error writing to %s", c->tempfile);
+		c = ap_proxy_cache_error(c);
             } else {
                 c->written += n;
             }
         }
 
 	/* Write the block to the client, detect aborted transfers */
-        while (n && !con->aborted) {
-            if (alt_to)
+        while (!con->aborted && n > 0) {
+            if (alternate_timeouts)
                 ap_soft_timeout("proxy send body", r);
 
             w = ap_bwrite(con->client, &buf[o], n);
 
-            if (alt_to)
+            if (alternate_timeouts)
                 ap_kill_timeout(r);
             else
                 ap_reset_timeout(r);
 
             if (w <= 0) {
-                if (c != NULL) {
+                if (c != NULL && c->fp != NULL) {
                     /* when a send failure occurs, we need to decide
                      * whether to continue loading and caching the
                      * document, or to abort the whole thing
                      */
                     ok = (c->len > 0) &&
                          (c->cache_completion > 0) &&
-                         (c->len * c->cache_completion < total_bytes_rcv);
+                         (c->len * c->cache_completion < total_bytes_rcvd);
 
                     if (! ok) {
-                        ap_pclosef(c->req->pool, c->fp->fd);
+                        ap_pclosef(c->req->pool, ap_bfileno(c->fp, B_WR));
                         c->fp = NULL;
                         unlink(c->tempfile);
 			c = NULL;
@@ -602,14 +612,14 @@ long int ap_proxy_send_fb(BUFF *f, request_rec *r, cache_req *c)
             }
             n -= w;
             o += w;
-        }
-    }
+        } /* while client alive and more data to send */
+    } /* loop and ap_bread while "ok" */
 
     if (!con->aborted)
 	ap_bflush(con->client);
 
     ap_kill_timeout(r);
-    return total_bytes_rcv;
+    return total_bytes_rcvd;
 }
 
 /*
@@ -625,14 +635,11 @@ void ap_proxy_send_headers(request_rec *r, const char *respline, table *t)
     BUFF *fp = r->connection->client;
     table_entry *elts = (table_entry *) ap_table_elts(t)->elts;
 
-    ap_bputs(respline, fp);
-    ap_bputs(CRLF, fp);
+    ap_bvputs(fp, respline, CRLF, NULL);
 
     for (i = 0; i < ap_table_elts(t)->nelts; ++i) {
 	if (elts[i].key != NULL) {
 	    ap_bvputs(fp, elts[i].key, ": ", elts[i].val, CRLF, NULL);
-	    /* FIXME: @@@ This used to be ap_table_set(), but I think
-	     * ap_table_addn() is correct. MnKr */
 	    ap_table_addn(r->headers_out, elts[i].key, elts[i].val);
 	}
     }
@@ -674,10 +681,10 @@ int ap_proxy_liststr(const char *list, const char *val)
     return 0;
 }
 
-#ifdef WIN32
+#ifdef CASE_BLIND_FILESYSTEM
 
 /*
- * On NT, the file system is NOT case sensitive. So, a == A
+ * On some platforms, the file system is NOT case sensitive. So, a == A
  * need to map to smaller set of characters
  */
 void ap_proxy_hash(const char *it, char *val, int ndepth, int nlength)
@@ -734,7 +741,7 @@ void ap_proxy_hash(const char *it, char *val, int ndepth, int nlength)
     char tmp[22];
     int i, k, d;
     unsigned int x;
-#if defined(AIX) && defined(__ps2__)
+#if defined(MPE) || (defined(AIX) && defined(__ps2__))
     /* Believe it or not, AIX 1.x does not allow you to name a file '@',
      * so hack around it in the encoding. */
     static const char enc_table[64] =
@@ -775,7 +782,7 @@ void ap_proxy_hash(const char *it, char *val, int ndepth, int nlength)
     val[i + 22 - k] = '\0';
 }
 
-#endif /* WIN32 */
+#endif /* CASE_BLIND_FILESYSTEM */
 
 /*
  * Converts 8 hex digits to a time integer
@@ -823,24 +830,34 @@ void ap_proxy_sec2hex(int t, char *y)
 
 cache_req *ap_proxy_cache_error(cache_req *c)
 {
-    ap_log_rerror(APLOG_MARK, APLOG_ERR, c->req,
-		 "proxy: error writing to cache file %s", c->tempfile);
-    ap_pclosef(c->req->pool, c->fp->fd);
-    c->fp = NULL;
-    unlink(c->tempfile);
+    if (c != NULL) {
+	if (c->fp != NULL) {
+	    ap_pclosef(c->req->pool, ap_bfileno(c->fp, B_WR));
+	    c->fp = NULL;
+	}
+	if (c->tempfile) unlink(c->tempfile);
+    }
     return NULL;
 }
 
-int ap_proxyerror(request_rec *r, const char *message)
+int ap_proxyerror(request_rec *r, int statuscode, const char *message)
 {
     ap_table_setn(r->notes, "error-notes",
 		  ap_pstrcat(r->pool, 
 			     "The proxy server could not handle the request "
-			     "<EM><A HREF=\"", r->uri, "\">",
-			     r->method, "&nbsp;", r->uri, "</A></EM>.<P>\n"
-			     "Reason: <STRONG>", message, "</STRONG>", NULL));
-    r->status_line = "500 Proxy Error";
-    return HTTP_INTERNAL_SERVER_ERROR;
+			     "<EM><A HREF=\"", ap_escape_uri(r->pool, r->uri),
+			     "\">", ap_escape_html(r->pool, r->method),
+			     "&nbsp;", 
+			     ap_escape_html(r->pool, r->uri), "</A></EM>.<P>\n"
+			     "Reason: <STRONG>",
+			     ap_escape_html(r->pool, message), 
+			     "</STRONG>", NULL));
+
+    /* Allow "error-notes" string to be printed by ap_send_error_response() */
+    ap_table_setn(r->notes, "verbose-error-to", ap_pstrdup(r->pool, "*"));
+
+    r->status_line = ap_psprintf(r->pool, "%3.3u Proxy Error", statuscode);
+    return statuscode;
 }
 
 /*
@@ -851,9 +868,7 @@ const char *
 {
     int i;
     struct hostent *hp;
-    static APACHE_TLS struct hostent hpbuf;
-    static APACHE_TLS u_long ipaddr;
-    static APACHE_TLS char *charpbuf[2];
+    struct per_thread_data *ptd = get_per_thread_data();
 
     for (i = 0; host[i] != '\0'; i++)
 	if (!ap_isdigit(host[i]) && host[i] != '.')
@@ -865,17 +880,17 @@ const char *
 	    return "Host not found";
     }
     else {
-	ipaddr = ap_inet_addr(host);
-	hp = gethostbyaddr((char *) &ipaddr, sizeof(u_long), AF_INET);
+	ptd->ipaddr = ap_inet_addr(host);
+	hp = gethostbyaddr((char *) &ptd->ipaddr, sizeof(ptd->ipaddr), AF_INET);
 	if (hp == NULL) {
-	    memset(&hpbuf, 0, sizeof(hpbuf));
-	    hpbuf.h_name = 0;
-	    hpbuf.h_addrtype = AF_INET;
-	    hpbuf.h_length = sizeof(u_long);
-	    hpbuf.h_addr_list = charpbuf;
-	    hpbuf.h_addr_list[0] = (char *) &ipaddr;
-	    hpbuf.h_addr_list[1] = 0;
-	    hp = &hpbuf;
+	    memset(&ptd->hpbuf, 0, sizeof(ptd->hpbuf));
+	    ptd->hpbuf.h_name = 0;
+	    ptd->hpbuf.h_addrtype = AF_INET;
+	    ptd->hpbuf.h_length = sizeof(ptd->ipaddr);
+	    ptd->hpbuf.h_addr_list = ptd->charpbuf;
+	    ptd->hpbuf.h_addr_list[0] = (char *) &ptd->ipaddr;
+	    ptd->hpbuf.h_addr_list[1] = 0;
+	    hp = &ptd->hpbuf;
 	}
     }
     *reqhp = *hp;
@@ -1230,7 +1245,7 @@ int ap_proxy_doconnect(int sock, struct sockaddr_in *addr, request_rec *r)
     ap_hard_timeout("proxy connect", r);
     do {
 	i = connect(sock, (struct sockaddr *) addr, sizeof(struct sockaddr_in));
-#ifdef WIN32
+#if defined(WIN32) || defined(NETWARE)
 	if (i == SOCKET_ERROR)
 	    errno = WSAGetLastError();
 #endif /* WIN32 */
@@ -1257,8 +1272,11 @@ int ap_proxy_send_hdr_line(void *p, const char *key, const char *value)
     if (!parm->req->assbackwards)
 	ap_rvputs(parm->req, key, ": ", value, CRLF, NULL);
     if (parm->cache != NULL && parm->cache->fp != NULL &&
-	ap_bvputs(parm->cache->fp, key, ": ", value, CRLF, NULL) == -1)
+	ap_bvputs(parm->cache->fp, key, ": ", value, CRLF, NULL) == -1) {
+	    ap_log_rerror(APLOG_MARK, APLOG_ERR, parm->cache->req,
+		    "proxy: error writing header to %s", parm->cache->tempfile);
 	    parm->cache = ap_proxy_cache_error(parm->cache);
+    }
     return 1; /* tell ap_table_do() to continue calling us for more headers */
 }
 
@@ -1271,3 +1289,44 @@ unsigned ap_proxy_bputs2(const char *data, BUFF *client, cache_req *cache)
     return len;
 }
 
+#if defined WIN32
+
+static DWORD tls_index;
+
+BOOL WINAPI DllMain (HINSTANCE dllhandle, DWORD reason, LPVOID reserved)
+{
+    LPVOID memptr;
+
+    switch (reason) {
+    case DLL_PROCESS_ATTACH:
+	tls_index = TlsAlloc();
+    case DLL_THREAD_ATTACH: /* intentional no break */
+	TlsSetValue (tls_index, malloc (sizeof (struct per_thread_data)));
+	break;
+    case DLL_THREAD_DETACH:
+	memptr = TlsGetValue (tls_index);
+	if (memptr) {
+	    free (memptr);
+	    TlsSetValue (tls_index, 0);
+	}
+	break;
+    }
+
+    return TRUE;
+}
+
+#endif
+
+static struct per_thread_data *get_per_thread_data(void)
+{
+#if defined(WIN32)
+
+    return (struct per_thread_data *) TlsGetValue (tls_index);
+
+#else
+
+    static APACHE_TLS struct per_thread_data sptd;
+    return &sptd;
+
+#endif
+}
